# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction",
        "description": "Documentation Auction web ",
        "contact": {
            "email": "rustamovvramziddin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Return only fields ",
                "description": "Get all data",
                "operationId": "d8d6b1e49fe70c44832e67a6c7134ff1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Post a new data",
                "description": "Post new user data",
                "operationId": "4e031d7222bb3a7a462abb6734194c81",
                "requestBody": {
                    "description": "Pass Product   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name",
                                    "image",
                                    "bidmargin",
                                    "description"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Product name "
                                    },
                                    "image": {
                                        "type": "binary",
                                        "format": "binary",
                                        "example": "image"
                                    },
                                    "bidmargin": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "$8000"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "This is a product description !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/product/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get one ",
                "description": "Return all date related to  product id}",
                "operationId": "587b2087957016512288cc57d5295d01",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Get one and Delete related to product id",
                "description": "Return  date related to ID of the Product",
                "operationId": "b1e3a3b1be8e5df0cda2cdfcc42b7834",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "put product id and try to delete ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/product/{id}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Product ",
                "description": "Returns updated Bid",
                "operationId": "Update Product price of the product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "bidding current update ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Update product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Product name "
                                    },
                                    "image": {
                                        "type": "binary",
                                        "format": "binary",
                                        "example": "image"
                                    },
                                    "bidmargin": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "$8000"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "This is a product description !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bid Updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "api": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}